% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyParseDEG.R
\name{DESeq_to_volcano_plot}
\alias{DESeq_to_volcano_plot}
\title{DESeq2 result to volcano plot}
\usage{
DESeq_to_volcano_plot(
  tab,
  pvalue.thres = 1,
  padj.thres = 0.05,
  log2fc.thres = 1,
  ylimit = c(0, 20),
  xlimit = c(-10, 10),
  point.size = 0.1,
  point.alpha = 0.1,
  volcano.title = "Gene expression change (treat vs ctrl)",
  yaxis = "padj",
  inc.color = "red",
  dec.color = "blue",
  ns.color = "grey60"
)
}
\arguments{
\item{tab}{differential expression table containing pvalue, padj and log2FoldChange.}

\item{pvalue.thres}{pvalue threshold for significant change.}

\item{padj.thres}{padj threshold for significant change.}

\item{log2fc.thres}{log2FC threshold for significant change.}

\item{ylimit}{volcano plot y axis range. default c(0,20).}

\item{xlimit}{volcano plot y axis range. default c(-10,10).}

\item{point.size}{point size. Default 0.1.}

\item{point.alpha}{point alpha. 0~1. Default 0.1.}

\item{volcano.title}{volcano plot title.}

\item{yaxis}{y axis type. padj or pvalue.}

\item{inc.color}{color for increased point. Default 'red'.}

\item{dec.color}{color for decreased point. Default 'blue'.}

\item{ns.color}{color for not significantly changed point. Dafault 'grey60'.}
}
\value{
a list containing table with change and volcano plot
}
\description{
provide a DESeq2 result, return a table with 'inc', 'dec', 'n.s.' state and a volcano plot.
}
\examples{
DESeq_to_volcano_plot(tab,pvalue.thres=1,padj.thres=0.05,log2fc.thres=1,ylimit=c(0,20),xlimit=c(-10,10),volcano.title='Gene expression change (treat vs ctrl)',yaxis='padj', inc.color='red',dec.color='blue',ns.color='grey60')
}
\author{
Shaobo Liang
}
