% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyDESeq.R
\name{EasyDESeq}
\alias{EasyDESeq}
\title{EasyDESeq}
\usage{
EasyDESeq(
  count_tab,
  tlabel,
  clabel,
  tRepN,
  cRepN,
  pthre = 0.05,
  foldchange = 1
)
}
\arguments{
\item{count_tab}{count table containing control and treatment samples. First control, second treatment.}

\item{tlabel}{label for treatment group}

\item{clabel}{label for control group}

\item{tRepN}{treatment replicates}

\item{cRepN}{control replicates}

\item{pthre}{padj threshold for summary output in console}

\item{foldchange}{fold change threshold for summary output in console}
}
\value{
a list containg DESeq2 result table and scale factor dataframe
}
\description{
EasyDESeq
}
\examples{
deseq(count_tab,'treatment','ctrl',3,3,0.05,2)
}
\author{
Shaobo Liang
}
