% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOdedup.R
\name{merge_similar_terms}
\alias{merge_similar_terms}
\title{merge similar GO term for dataframe output from gprofiler2}
\usage{
merge_similar_terms(gostres.tab)
}
\arguments{
\item{gostres.tab}{result table from gprofiler2::gost. intersection column is required in dataframe. intersection is gene list.}
}
\value{
a subset dataframe of input
}
\description{
merge similar GO term for dataframe output from gprofiler2
}
\details{
none
}
\examples{
merge_similar_terms(gostres)
}
\author{
Shaobo Liang
}
