% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM.R
\name{readcount_to_TPM}
\alias{readcount_to_TPM}
\title{Title}
\usage{
readcount_to_TPM(gene.readcount.tab)
}
\arguments{
\item{gene.readcount.tab}{readcount table contains at 3 columns. Geneid, Length, and other sample count columns, like output from featureCounts}
}
\value{
TPM matrix. rownames are gene ids
}
\description{
transfer readcount matrix into TPM
}
