% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EasyParseDAS.R
\name{rmats_to_volcano_plot}
\alias{rmats_to_volcano_plot}
\title{rmats result to volcano plot}
\usage{
rmats_to_volcano_plot(
  tab,
  PValue.thres = 1,
  FDR.thres = 0.05,
  deltaPSI.thres = 0.05,
  ylimit = c(0, 10),
  xlimit = c(-1, 1),
  volcano.title = "PSI change (treat vs ctrl)",
  yaxis = "FDR",
  inc.color = "red",
  dec.color = "blue",
  ns.color = "grey60"
)
}
\arguments{
\item{PValue.thres}{PValue threshold for significant change.}

\item{FDR.thres}{FDR threshold for significant change.}

\item{deltaPSI.thres}{delta PSI threshold for significant change.}

\item{ylimit}{volcano plot y axis range. default c(0,10).}

\item{xlimit}{volcano plot y axis range. default c(-1,1).}

\item{volcano.title}{volcano plot title.}

\item{yaxis}{y axis type. PValue or FDR.}

\item{inc.color}{color for increased point. Default 'red'.}

\item{dec.color}{color for decreased point. Default 'blue'.}

\item{ns.color}{color for not significantly changed point. Dafault 'grey60'.}

\item{DAtab}{Differentially alternative splicing table. PValue, FDR, deltaPSI (IncLevelDifference) are required.}
}
\value{
a list containing table with change and volcano plot, violin plot
}
\description{
rmats result to volcano plot
}
\examples{
rmats_to_volcano_plot(tab,PValue.thres=1,FDR.thres=0.05,deltaPSI.thres=0.05,ylimit=c(0,10),xlimit=c(-1,1),volcano.title='Differentially AS (treat vs ctrl)',yaxis='FDR', inc.color='red',dec.color='blue',ns.color='grey60')
}
